Test: ApiTest에 AdminApiTest 통합
