+++ src/main/java/com/keunsori/keunsoriserver/domain/admin/account/service/AdminAccountService.java
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class AdminAccountService {
 
-    private final MemberRepository memberRepository;
     private final MemberUtil memberUtil;
 
     public AdminMyPageResponse getAdminMyPage() {
