+++ src/test/java/com/keunsori/keunsoriserver/common/ApiTest.java
@@ -18,10 +18,6 @@ import org.springframework.test.context.ActiveProfiles;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.keunsori.keunsoriserver.domain.auth.login.dto.request.LoginRequest;
-import com.keunsori.keunsoriserver.domain.member.domain.Member;
-import com.keunsori.keunsoriserver.domain.member.domain.vo.MemberStatus;
-import com.keunsori.keunsoriserver.domain.member.dto.request.SignUpRequest;
-import com.keunsori.keunsoriserver.domain.member.repository.MemberRepository;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
@@ -39,7 +35,9 @@ public class ApiTest {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-    protected String token;
+    protected String generalToken;
+
+    protected String adminToken;
 
     @AfterEach
     public void clear() {
@@ -59,7 +57,7 @@ public class ApiTest {
 
         LoginRequest request = new LoginRequest("C000001", "test123!");
 
-        token = given().
+        generalToken = given().
                         header(CONTENT_TYPE, "application/json").
                         body(mapper.writeValueAsString(request)).
                 when().
@@ -68,5 +66,29 @@ public class ApiTest {
                         statusCode(SC_OK).
                         extract().
                         jsonPath().getString("accessToken");
+    }
+
+    @Test
+    public void login_with_admin_member() throws JsonProcessingException {
+        memberRepository.deleteAll();
+        Member member = Member.builder()
+                .studentId("A000001")
+                .email("testAdmin@g.hongik.ac.kr")
+                .password(passwordEncoder.encode("testadmin123!"))
+                .status(MemberStatus.관리자)
+                .build();
+        memberRepository.save(member);
+
+        LoginRequest request = new LoginRequest("A000001", "testadmin123!");
+
+        adminToken = given().
+                header(CONTENT_TYPE, "application/json").
+                body(mapper.writeValueAsString(request)).
+                when().
+                post("/auth/login").
+                then().
+                statusCode(SC_OK).
+                extract().
+                jsonPath().getString("accessToken");
     }
 }
