+++ src/test/java/com/keunsori/keunsoriserver/reservation/api/ReservationApiTest.java
@@ -39,7 +39,7 @@ public class ReservationApiTest extends ApiTest {
     @BeforeEach
     void login() throws JsonProcessingException {
         login_with_general_member();
-        authorizationValue = "Bearer " + token;
+        authorizationValue = "Bearer " + generalToken;
     }
 
     @Test
